make it possible to reference fields of enum variant in assoc
Field values of an enum variant can be referenced in assoc attribute expressions by their names prefixed with an underscore. The underscore is there to make linter happy if the value is not used. In case of tuples the values are available under "_0", "_1", and so on.

The tests for this feature are minimal.